Support drives > 8.4GB
For LBA-capable drives, return the number of LBA sectors as secperunit, not
the number you get from CHS.

Note, DOS partition tables are limited to 8.4GB. The interactive
disklabel editor will notice that the OpenBSD partition ends at 8.4GB and
limit the user to that size unless the user explicitly specifies the bounds.
This should probably be changed.