Better key repeat logic. When we're in repeat mode, with the exception of modifier keys (SHIFT, CTRL, ALT, etc.), and a key "up" event is received, only delete the key repeat timeout if the key that generated the "up" event is the key currently being repeated. When a key "down" event is received, the opposite -- only delete the key repeat if the key that generated the "down" event is _not_ the key currently being repeated.
Playing here is a bit dangerous since we are constraining the conditions in
which a timeout(9) is deleted (we don't want to mistakenly get stuck in a
non-legitamite key repeat).

Using timeout(9) for this is kind of gross. I guess it is done here to avoid
key repeat code in all of the actual hardware keyboard drivers.

There is still a bug here. If you have a key held down and detach your
keyboard (by either unplugging it or doing a KVM switch), the timeout is not
deleted and the key continues to repeat until it gets the keyboard back and
receives an interrupt. Perhaps we should not be ignoring typematic keys after
all?