Nadav Shemer of Tehuti Networks is magical.
not leaving the gap in the txt fifo when uploading the firmware was the
cause of my fifo write bug in tht_start. because i was filling the whole
fifo, i was writing the wptr back to its original position. because of
this i dont think the firmware thought i had written anything. only the
last short chunk would have been noticed, which strikes me as possibly
confusing to the chip.

this diff removes the delay at the top of tht_fifo_post. one less XXX :)