Increase the size of the diskinfo struct, which has grown 4 bytes in the new firmware.
To allow this to work on older firmwares add a "short read allowed"
flag to arc_msgbuf(), which will allow short reads by getting
the read length from the header of the response the firmware sends.

Only set the flag on the arc_msgbuf() call that deals with the diskinfo
struct for now.

Discussed with and ok dlg@