Rework the vblank subsystem so that instead of having various bits in the drm softc, we only have one pointer, with the rest in that struct. This is so that vblank-less drivers (yes, they exist) don't need to waste space on useless crud.
While i'm reworking most of this code anyway, accept that on openbsd all
of the #defined atomic functions are not atomic other than set and clear
bit. Also, realise that the vb_lock is held whenever we manipulate these
counts anyway.  With those two facts in mind just remove the
atomic_blah() and just use ++ and --.