Define struct cpu_hwinfo, to hold hardware specific information about each processor (instead of sys_config.cpu[]), and pass it in the attach_args when attaching cpu devices.
This allows per-cpu information to be gathered late in the bootstrap process,
and not be limited by an arbitrary MAX_CPUS limit; this will suit IP27 and
IP35 systems better.

While there, use this information to make sure delay() uses the speed
information from the cpu it is invoked on.