Prevent a possible case of lock recursion in swapoff.
If when we have successfully swapped an aobj back in, then we release our
reference count, and that reference is the last reference, we will free the
the aobj and recursively lock the list lock.

Fix this by keeping track of the last object we had a reference on, and
releasing the refcount the next time we unlock the list lock.

Put a couple of comments in explaining lock ordering in this file.

noticed by, discussed with and ok guenther@.