amd64_pa_used is weird.
We check a bunch of things, where most archs are a lot more simple.
Also, we get problems sometimes: My x201 can't map the framebuffer BAR
from the aperture with the chunk that checks the bios sections, meaning
the bios is claiming something about it.

Kettenis@ and I are pretty sure that the rest of the checks are more than
sufficient, so just nuke that one.

for the record, in the same situation i386 just checks again VGA_START,
BIOS_END and physmem.

ok kettenis@