Checksum fix from reyk (rev1.31) that was lost during update:
Fix the IP ckecksum offloading logic that disables and breaks offloading
if the packet is neither TCP nor UDP because of an erroneous "default" case.
No functional change in the default build because IP checksum offloading
is currently disabled in ix(4).

Tested on 82598 and 82599 (though checksums are still disabled by default).

ok claudio