In in6_update_ifa(), don't set the TENTATIVE flag on an IPv6 address that is marked as NODAD.
Since we're not gonna do duplicate address detection for this address,
the TENTATIVE flag won't get cleared, rendering the address unusable.
The existing logic only checked whether DAD was enabled for the interface,
but DAD can be configured on a per-address basis.

ok sthen@ as part of a larger diff