provide an MI api for doing byteswapping loads and stores. some archs have instrutions that can do this, and the rest that dont get to use wrappers around the byteswap(3) api.
this provides MI backends for sparc64 and powerpc which get a big
benefit from this because byteswapping in registers is really hard
for them.

the intended use case is for reading and writing bits of dma memory
handed to and from hardware.

discussed with miod@ guenther@ deraadt@
ok miod@ kettenis@