Correctly recognize Super Speed devices, this is part of the work to be able to use USB 3.0 devices behind an external hub.
This is a bit tricky because the SS status use a different power bit
that maps to the Low speed one.  So no longer accept devices without
power bit and fallback to the parent hub's speed in case the status
does not report any particular speed.

Note that xhci(4) root hubs still set the traditionnal UPS_PORT_POWER
bit with the correct device speed.