Update inteldrm to the code from Linux 3.14.52 (which corresponds to commit 48f8f36a6c8018c2b36ea207aaf68ef5326c5075 on the linux-3.14.y branch of the linux-stable tree).  This brings preliminary support for the GPU on Intel's Broadwell CPUs.  Don't expect these to work perfectly yet.  There are some remaining issues with older hardware as well, but no significant regressions have been uncovered.
This also updates some of drm core code.  The radeondrm code remains
based on Linux 3.8 with some minimal canges to adjust to changes in
the core drm APIs.

Joint effort with jsg@, who did the initial update of the relevant drm
core bits.  Committing this early to make sure it gets more testing
and make it possible for others to help getting the remaining wrinkles
straightened out.